2018_07_02
.

+++ 최규화/Mod_Server/Server/CPhysicEngineWorker.h
@@ -4,8 +4,8 @@ class CPhysicEngineWorker
 private:
 	high_resolution_clock::time_point	m_currtime, m_prevtime;
 
-	 
 	float m_deltime;
+	
 
 	high_resolution_clock::time_point tt;
 
@@ -14,7 +14,7 @@ public:
 	void CheckCurrTime() { m_currtime = high_resolution_clock::now(); }
 	void CheckPrevTime() { m_prevtime = high_resolution_clock::now(); }
 	void Update();
-	void CollisionSystem(float deltime);
+	void CollisionSystem(__int64 deltime);
 	~CPhysicEngineWorker();
 };
